requirements.txt
fpdf2==2.5.6
Pillow==9.2.0
pylibdmtx==0.1.10
tk==0.1.0
